agents-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.1.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==3.7.2
certifi==2022.9.24
cffi==2.0.0
charset-normalizer==2.1.1
click==8.2.0
colorama==0.4.6
comm==0.1.1
contourpy==1.3.1
cryptography==46.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.0
executing==1.2.0
fastapi==0.115.12
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
greenlet==3.2.4
griffe==1.7.3
h11==0.14.0
h2==4.3.0
h5py==3.7.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.4
ipykernel==6.18.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.9.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_bokeh==4.0.5
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.27.3
kiwisolver==1.4.4
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mcp==1.9.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==2.0.4
multidict==6.7.0
mysql-connector-python==8.0.33
narwhals==1.34.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==7.3.3
notebook_shim==0.2.2
numpy==1.23.5
openai==1.109.1
openai-agents==0.0.15
opencv-python==4.6.0.66
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
panel==1.6.2
param==2.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
postgrest==2.22.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
propcache==0.4.1
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.13.0
PyJWT==2.10.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyviz_comms==3.0.4
pywin32==305
pywinpty==2.0.9
PyYAML==6.0.2
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
realtime==2.5.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.24.0
Send2Trash==1.8.3
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.43
sse-starlette==2.3.5
stack-data==0.6.2
starlette==0.46.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tenacity==9.1.2
terminado==0.17.0
tiktoken==0.11.0
tinycss2==1.2.1
tomli==2.2.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.4.0
utils==1.0.2
uvicorn==0.34.2
watchfiles==1.1.1
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.3
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.22.0
zstandard==0.25.0
